updating yml tests to run as command
