+++ Gemfile
@@ -2,9 +2,13 @@ source 'https://rubygems.org'
 
 gem 'jekyll', '~> 3.8.2'
 gem 'minimal-mistakes-jekyll', '4.9.0'
-gem 'jekyll-admin', group: :jekyll_plugins
 
 gem 'wdm', '~> 0.1.0' if Gem.win_platform?
 gem 'html-proofer', '3.9.1'
 gem 'mdl', '0.4.0'
 gem 'jekyll-responsive-image', '1.2.0'
+
+group :jekyll_plugins do
+  gem 'jekyll-admin'
+  gem 'jekyll-frontmatter-test', '0.0.1', :path => '_plugins'
+end
