+++ _plugins/lib/jekyll-frontmatter-test.rb
@@ -1,0 +1,7 @@
+module Jekyll
+  module FrontmatterTest
+    autoload :VERSION, 'jekyll/frontmatter-test/version.rb'
+  end
+end
+
+require 'jekyll/commands/frontmatter-test.rb'
