+++ _plugins/lib/jekyll/frontmatter-test/version.rb
@@ -1,0 +1,5 @@
+module Jekyll
+  module FrontmatterTest
+    VERSION = '0.0.1'
+  end
+end
